+++ config/event-projector.php
@@ -48,6 +48,13 @@ return [
     'projector_status_model' => \Spatie\EventProjector\Models\ProjectorStatus::class,
 
     /*
+     * This class is responsible for handle stored events. To add extra behaviour you
+     * can change this to a class of your own. The only restriction is that
+     * it should extend \Spatie\EventProjector\HandleStoredEventJob.
+     */
+    'stored_event_job' => \Spatie\EventProjector\HandleStoredEventJob::class,
+
+    /*
      * This class is responsible for serializing events. By default an event will be serialized
      * and stored as json. You can customize the class name. A valid serializer
      * should implement Spatie\EventProjector\EventSerializers\Serializer.
@@ -57,7 +64,7 @@ return [
     /*
      * When replaying events potentially a lot of events will have to be retrieved.
      * In order to avoid memory problems events will be retrieved in
-     * a chuncked way. You can specify the chunk size here.
+     * a chunked way. You can specify the chunk size here.
      */
     'replay_chunk_size' => 1000,
 ];