+++ src/EventHandlers/EventHandlerCollection.php
@@ -4,6 +4,7 @@ namespace Spatie\EventProjector\EventHandlers;
 
 use Illuminate\Support\Collection;
 use Spatie\EventProjector\Models\StoredEvent;
+use Spatie\EventProjector\Exceptions\InvalidEventHandler;
 
 class EventHandlerCollection
 {